requirements.txt
Django==4.1.1
psycopg2==2.9.5
scikit-learn==1.2.2
pandas==2.0.0
joblib==1.2.0
Pillow==9.5.0